# app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "iSECTECH Notification & Trust Score API",
    "description": "Production-grade API for notification management and trust scoring systems supporting mobile PWA and enterprise security operations",
    "version": "1.0.0",
    "contact": {
      "name": "iSECTECH API Team",
      "email": "api@isectech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://isectech.com/license"
    }
  },
  "servers": [
    {
      "url": "https://api.isectech.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://api-staging.isectech.com/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/notifications/preferences": {
      "get": {
        "tags": ["Notification Preferences"],
        "summary": "Get user notification preferences",
        "description": "Retrieve notification preferences for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "User ID (optional, defaults to authenticated user)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preferences retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationPreferences"
                    },
                    "metadata": {
                      "$ref": "#/components/schemas/ResponseMetadata"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": ["Notification Preferences"],
        "summary": "Update user notification preferences",
        "description": "Update notification preferences for a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "User ID (optional, defaults to authenticated user)"
                  },
                  "preferences": {
                    "$ref": "#/components/schemas/NotificationPreferencesUpdate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationPreferences"
                    },
                    "metadata": {
                      "$ref": "#/components/schemas/ResponseMetadata"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "NotificationPreferences": {
        "type": "object",
        "properties": {
          "channels": {
            "type": "object",
            "properties": {
              "push": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "quietHours": {
                    "type": "object",
                    "properties": {
                      "enabled": {
                        "type": "boolean"
                      },
                      "startTime": {
                        "type": "string",
                        "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
                      },
                      "endTime": {
                        "type": "string",
                        "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
                      },
                      "timezone": {
                        "type": "string"
                      }
                    }
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["security", "alert", "info", "warning", "error"]
                    }
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "frequency": {
                    "type": "string",
                    "enum": ["immediate", "hourly", "daily", "weekly"]
                  },
                  "digest": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "globalSettings": {
            "type": "object",
            "properties": {
              "doNotDisturb": {
                "type": "boolean"
              },
              "language": {
                "type": "string",
                "enum": ["en", "es", "fr", "de", "ja", "zh"]
              },
              "timezone": {
                "type": "string"
              }
            }
          }
        }
      },
      "NotificationPreferencesUpdate": {
        "type": "object",
        "properties": {
          "channels": {
            "$ref": "#/components/schemas/NotificationPreferences/properties/channels"
          },
          "globalSettings": {
            "$ref": "#/components/schemas/NotificationPreferences/properties/globalSettings"
          }
        }
      },
...
      "ResponseMetadata": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "req_123456789"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "processingTime": {
            "type": "integer",
            "description": "Processing time in milliseconds",
            "example": 150
          },
          "cached": {
            "type": "boolean",
            "description": "Whether the response was served from cache"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid query parameters"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                },
                "requestId": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal server error"
                },
                "requestId": {
                  "type": "string"
                },
                "details": {
                  "type": "string",
                  "description": "Error details (only in development mode)"
                }
              }
            }
          }
        }
      }
    }
  }
}